# Performance: improve parsing of moments from valid Date objects