--- a/src/lib/create/from-array.js
+++ b/src/lib/create/from-array.js
@@ -5,6 +5,7 @@ import { dayOfYearFromWeeks } from '../units/day-of-year';
 import { YEAR, MONTH, DATE, HOUR, MINUTE, SECOND, MILLISECOND } from '../units/constants';
 import { createLocal } from './local';
 import defaults from '../utils/defaults';
+import getParsingFlags from './parsing-flags';
 
 function currentDateArray(config) {
     var now = new Date();
@@ -37,7 +38,7 @@ export function configFromArray (config) {
         yearToUse = defaults(config._a[YEAR], currentDate[YEAR]);
 
         if (config._dayOfYear > daysInYear(yearToUse)) {
-            config._pf._overflowDayOfYear = true;
+            getParsingFlags(config)._overflowDayOfYear = true;
         }
 
         date = createUTCDate(yearToUse, 0, config._dayOfYear);