--- a/src/lib/moment/valid.js
+++ b/src/lib/moment/valid.js
@@ -1,14 +1,15 @@
 import { isValid as _isValid } from '../create/valid';
 import extend from '../utils/extend';
+import getParsingFlags from '../create/parsing-flags';
 
 export function isValid () {
     return _isValid(this);
 }
 
 export function parsingFlags () {
-    return extend({}, this._pf);
+    return extend({}, getParsingFlags(this));
 }
 
 export function invalidAt () {
-    return this._pf.overflow;
+    return getParsingFlags(this).overflow;
 }