--- a/src/lib/units/hour.js
+++ b/src/lib/units/hour.js
@@ -5,6 +5,7 @@ import { addRegexToken, match1to2, match2 } from '../parse/regex';
 import { addParseToken } from '../parse/token';
 import { HOUR } from './constants';
 import toInt from '../utils/to-int';
+import getParsingFlags from '../create/parsing-flags';
 
 // FORMATTING
 
@@ -46,7 +47,7 @@ addParseToken(['a', 'A'], function (input, array, config) {
 });
 addParseToken(['h', 'hh'], function (input, array, config) {
     array[HOUR] = toInt(input);
-    config._pf.bigHour = true;
+    getParsingFlags(config).bigHour = true;
 });
 
 // LOCALES