--- a/src/lib/utils/is-date.js
+++ b/src/lib/utils/is-date.js
@@ -1,3 +1,3 @@
 export default function isDate(input) {
-    return Object.prototype.toString.call(input) === '[object Date]' || input instanceof Date;
+    return input instanceof Date || Object.prototype.toString.call(input) === '[object Date]';
 }