EDIT: es6 stuff is merged, so this PR needs to get rewritten.  I do plan on getting to it, but not immediately.

For explanation of some of these changes, I'll discuss my numbers.  I'm doing this from memory and can re-measure if necessary, but...  My CPU profile in Chrome started at 400ms, and after this suite of changes, the profile was reduced to 200ms.

* https://github.com/brianwyantora/moment/commit/6bc5d2e075b0cad32083ea5209bb6ac695029aa7
  - This change resulted in a reduction of about 70ms on my profile.
  - Basically, in my use case, I never send invalid dates from server to client.  So while I might do some work on the first date sent from the server to make sure the browser can parse it, once I've done that, I won't call `moment.fn.isValid()`.  Generating the parsing flags takes a notable amount of time, so deferring that generation is a significant help.
* https://github.com/brianwyantora/moment/commit/928e068349c10d90bceb5266727e016f0cda1259
  - This change resulted in a reduction of about 60ms on my profile.
  - This change might look wrong on first glance, but notice - `makeMoment` is only called from two places (the public constructors), and will pass the config object through `new Moment(config)`.  `Moment` already clones the date, so doing the same in `makeMoment` is just duplicating work unnecessarily.