@icambron 2015-04-04
I don't have any feedback on the changes themselves, but I wanted to suggest that you add some benchmarking to the product itself. Just add a file here: https://github.com/moment/moment/tree/develop/benchmarks