@brianwyantora 2015-04-06
Sure.  Don't suppose you have any setup already for using browserify with those benchmark files, or something, to get them running in the browser?

Anyway, running *grunt benchmark* gave the following:

*Before*: 685,699 ops/sec ±0.84%
*After*: 1,748,813 ops/sec ±0.57%