@icambron 2015-04-06
Thanks. No, I don't have a good way to run those perf tests in the browsers, though I think it's fair to say that if it's faster in Node, it'll be faster in Chrome.