@ichernev 2015-04-17
@brianwyantora looks pretty good. I'll merge it once you port it to the ES6 modules, sorry for that :-/