@brianwyantora 2015-04-22
ES6 before changes:
```
Running benchmark clone [benchmarks/clone.js]...
>> clone x 3,318,252 ops/sec ±1.20% (87 runs sampled)
Running benchmark fromDate [benchmarks/fromDate.js]...
>> clone x 571,146 ops/sec ±0.81% (85 runs sampled)
Running benchmark fromDateUtc [benchmarks/fromDateUtc.js]...
>> clone x 549,873 ops/sec ±0.82% (89 runs sampled)
```

ES6 after changes:
```
Running benchmark clone [benchmarks/clone.js]...
>> clone x 3,856,373 ops/sec ±0.70% (91 runs sampled)
Running benchmark fromDate [benchmarks/fromDate.js]...
>> clone x 1,682,164 ops/sec ±1.50% (92 runs sampled)
Running benchmark fromDateUtc [benchmarks/fromDateUtc.js]...
>> clone x 1,717,440 ops/sec ±0.94% (89 runs sampled)
```

Ran `grunt test update-index benchmark` for the after benchmarks.  The big whopper of a method now seems to be `createFromConfig`, which I've left untouched.  Please review carefully, I wasn't as critical of my changes as I was earlier.