@ichernev 2015-05-04
Merged in 3f65e56c3d04a5c0d8c35fd4a3008af15f5ec1e2